feat(embeddings): add cnn embedding
